Use custom ApplyTransforms.
